+++ runner.js
@@ -1,99 +1,27 @@
-// module: Queueman.Runner
-let fivebeans = require('fivebeans');
-let request = require('request-promise');
+const fivebeans = require('fivebeans');
+const request = require('request-promise');
 
-module.exports = function (handle, errorHandler = (() => {}), host = '127.0.0.1', port = 11300) {
-
-	function Runner(handle, errorHandler, host, port) {
+class Runner {
+	constructor(
+		handle,
+		errorHandler = (() => {}),
+		host = '127.0.0.1',
+		port = 11300,
+	) {
+		this.handle = handle;
 		this.errorHandler = errorHandler;
 		this.host = host;
 		this.port = port;
-
-		const self = this;
-
-		// Builds the handler used by the queue worker
-		this.handler = function () {
-			function MainJobHandler() {
-				this.type = 'mainjob';
-			}
-
-			// Called within the fivebeans package
-			MainJobHandler.prototype.work = function (payload, callback) {
-				handle(payload).then(products => {
-					callback('success');
-
-					console.log((new Date).toUTCString() + " Found a job!");
-
-					console.log((new Date).toUTCString() + " Found " + products.length + " products");
-
-					if (payload.callback_url) {
-						// If there is a callback_url send the deal_data back to it.
-						let options = {
-							uri: payload.callback_url,
-							method: 'POST',
-							body: {
-								params: {},
-								products,
-							},
-							json: true,
-						};
-						console.log((new Date).toUTCString() + " Sending request to: " + payload.callback_url);
-						request.post(options).then(body => {
-								console.log((new Date).toUTCString() + " Job complete!");
-							},
-							err => {
-								console.log((new Date).toUTCString() + " Failed to send products to " + payload.callback_url);
-								self.errorHandler(err);
-							});
-					}else{
-						// No callback_url so just succeed.
-						console.log((new Date).toUTCString() + " Job complete!");
-					}
-
-				}, err => {
-					callback('success');
-					console.log((new Date).toUTCString() + " Job failed, burying job!");
-					self.errorHandler(err);
-					// If we have a callback_url send the error
-					if (payload.callback_url) {
-						if (typeof err === 'object' && err.message) {
-							err = err.message;
-						}
-
-						let options = {
-							uri: payload.callback_url,
-							method: 'POST',
-							body: {
-								error: err,
-							},
-							json: true,
-						};
-						request.post(options).then(() => {
-							console.log((new Date).toUTCString() + " Sent error to "+payload.callback_url);
-						}, err => {
-							console.log((new Date).toUTCString() + " Failed to send error to "+payload.callback_url);
-							self.errorHandler(err);
-						})
-					}
-				});
-			}
-			let handler = new MainJobHandler();
-			return handler;
-		}
 	}
 
-	// Starts the runner and attaches the handler
-	Runner.prototype.run = function () {
-		let options = {
+	run() {
+		const worker = new fivebeans.worker({
 			id: 'worker_1',
 			host: this.host,
 			port: this.port,
-			handlers: {
-				mainjob: this.handler()
-			},
+			handlers: { mainjob: this },
 			ignoreDefault: false,
-		}
-		let worker = new fivebeans.worker(options);
+		});
 
 		worker.on('info', (info) => {
 			console.log('==== INFO ====');
@@ -118,6 +46,86 @@ module.exports = function (handle, errorHandler = (() => {}), host = '127.0.0.1'
 		worker.start();
 	}
 
-	// Return the class
-	return new Runner(handle, errorHandler, host, port);
+	work(payload, callback) {
+		const callbackUrl = payload.callback_url;
+
+		if (callbackUrl) {
+			this.sendStartedCallback(callbackUrl);
+		}
+
+		this.handle(payload)
+			.then((products) => {
+				callback('success');
+
+				this.log('Found a job!');
+				this.log(`Found ${products.length} products`);
+
+				if (callbackUrl) {
+					this.sendResultCallback(callbackUrl, products);
+				}
+
+				this.log('Job complete!');
+			})
+			.catch((error) => {
+				callback('success');
+
+				this.errorHandler(error);
+				this.log('Job failed, burying job!');
+
+				if (callbackUrl) {
+					this.sendErrorCallback(callbackUrl, error);
+				}
+			});
+	}
+
+	sendStartedCallback(url) {
+		this.log(`Sending started callback to ${url}`);
+
+		return this.sendCallback(url, { started_at: (new Date).toISOString() })
+			.then(() => this.log('Sent started callback successfully'))
+			.catch((error) => {
+				this.errorHandler(error);
+				this.log(`Failed to send started callback to ${url}`);
+			});
+	}
+
+	sendResultCallback(url, products) {
+		this.log(`Sending result callback to ${url}`);
+
+		this.sendCallback(url, { products })
+			.catch((error) => {
+				this.errorHandler(error);
+				this.log(`Failed to send products to ${url}`);
+			});
+	}
+
+	sendErrorCallback(url, error) {
+		this.log(`Sending error callback to ${url}`);
+
+		if (typeof error === 'object' && error.message) {
+			error = error.message;
+		}
+
+		this.sendCallback(url, { error })
+			.then(() => this.log(`Sent error to ${url}`))
+			.catch((error) => {
+				this.errorHandler(error);
+				this.log(`Failed to send error to ${url}`);
+			});
+	}
+
+	sendCallback(url, data) {
+		return request.post({
+			uri: url,
+			method: 'POST',
+			body: data,
+			json: true,
+		});
+	}
+
+	log(message) {
+		console.log(`${(new Date).toUTCString()} ${message}`);
+	}
 }
+
+module.exports = Runner;