+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2018-07-27
+
+### Added
+
+- Added started callback ([#375])
+
+### Changed
+
+- Refactored Runner
+
 ## [1.2.1] - 2018-06-20
 
 ### Fixed
@@ -57,7 +67,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.1.1]: https://bitbucket.org/razorcreations/feedme-web-refresh/branches/compare/1.1.1..1.1.0
 [1.2.0]: https://bitbucket.org/razorcreations/feedme-web-refresh/branches/compare/1.2.0..1.1.1
 [1.2.1]: https://bitbucket.org/razorcreations/feedme-web-refresh/branches/compare/1.2.1..1.2.0
+[1.3.0]: https://bitbucket.org/razorcreations/feedme-web-refresh/branches/compare/1.3.0..1.2.1
 
+[#375]: https://app.activecollab.com/160962/projects/18/tasks/1207
 [#392]: https://app.activecollab.com/160962/projects/18/tasks/1248
 [#417]: https://app.activecollab.com/160962/projects/18/tasks/1327
 [#447]: https://app.activecollab.com/160962/projects/18/tasks/1373